s_product_template_multi_link: split mixin to its own file
