+++ shopinvader_product_template_multi_link/models/__init__.py
@@ -1,1 +1,2 @@
+from . import shopinvader_product_link_mixin
 from . import shopinvader_product
